addressed comment 2

+++ glide-core/telemetry/src/lib.rs
@@ -29,7 +29,7 @@ pub struct Telemetry {
     /// Number of times compression was skipped
     compression_skipped_count: usize,
     /// Number of times subscriptions were detected as out of sync
-    subscription_out_of_sync_count: u64,
+    subscription_out_of_sync_count: usize,
     /// Unix timestamp (in milliseconds) of the last time subscriptions were in sync
     subscription_last_sync_timestamp: u64,
 }
@@ -172,14 +172,14 @@ impl Telemetry {
     }
     /// Increment the subscription out of sync count
     /// Return the new count after increment
-    pub fn incr_subscription_out_of_sync() -> u64 {
+    pub fn incr_subscription_out_of_sync() -> usize {
         let mut t = TELEMETRY.write().expect(MUTEX_WRITE_ERR);
         t.subscription_out_of_sync_count = t.subscription_out_of_sync_count.saturating_add(1);
         t.subscription_out_of_sync_count
     }
 
     /// Get the current subscription out of sync count
-    pub fn subscription_out_of_sync_count() -> u64 {
+    pub fn subscription_out_of_sync_count() -> usize {
         TELEMETRY
             .read()
             .expect(MUTEX_READ_ERR)
